requirements.txt
django
django-rest-framework
django-cors-headers
requests
lxml
bs4
celery
redis
sqlalchemy
psycopg2-binary